% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_paths_file.R
\name{import_paths_file}
\alias{import_paths_file}
\title{Import provided paths file}
\usage{
import_paths_file(
  paths_file_path,
  check_file_table_headers = T,
  col_types = "ccccccc"
)
}
\arguments{
\item{paths_file_path}{path to the paths_file}

\item{check_file_table_headers}{logical value that whether the path file should be checked for containing correct headers. Defaults to TRUE}

\item{col_types}{overwrite the column types of the imported paths file.}
}
\value{
tibble with the provided paths of all samples
}
\description{
Import provided paths file
}
