% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_chunks.R
\name{merge_Rds_chunks}
\alias{merge_Rds_chunks}
\title{Merge RDS chunks containing data frames or data tables into one single RDS file}
\usage{
merge_Rds_chunks(
  pattern,
  folder,
  output_file_path,
  verbose = FALSE,
  max_size_all_chunks = 2e+09
)
}
\arguments{
\item{pattern}{pattern to match the RDS chunk files}

\item{folder}{folder to search for RDS chunk files}

\item{output_file_path}{where to store the merged output file (full_path)}

\item{verbose}{should verbose logging be activated? (TRUE / FALSE)}

\item{max_size_all_chunks}{the maximum size of all the chunks combined that would still be merged}
}
\description{
Merge RDS chunks containing data frames or data tables into one single RDS file
}
