% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_gene_annotation_ref_file.R
\name{prepare_gene_annotation_ref_file}
\alias{prepare_gene_annotation_ref_file}
\title{Prepare the gene annotation reference files from default online sources}
\usage{
prepare_gene_annotation_ref_file(
  gencode_gtf_file_path,
  imprint_genes_file_path,
  cancer_gene_file_path,
  ref_file_output_path,
  ref_file_output_path_exons
)
}
\arguments{
\item{gencode_gtf_file_path}{Path to the Gencode GTF file. Please download from https://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_40/GRCh37_mapping/gencode.v40lift37.annotation.gtf.gz}

\item{imprint_genes_file_path}{Path to the Gencode GTF file. Please download from https://www.geneimprint.com/site/genes-by-species or SOURCE}

\item{cancer_gene_file_path}{Path to the cancer gene census file. Please download from https://cancer.sanger.ac.uk/census or SOURCE}

\item{ref_file_output_path}{Output path for the gene annotion reference file as required for Revana}

\item{ref_file_output_path_exons}{Output path for the exon annotion reference file as required for Revana}
}
\description{
Prepare the gene annotation reference files from default online sources
}
