% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_genehancer_ref_file.R
\name{prepare_genehancer_ref_file}
\alias{prepare_genehancer_ref_file}
\title{Prepare the GeneHancer reference file}
\usage{
prepare_genehancer_ref_file(
  genehancer_data,
  output_path,
  genehancer_element_elite_status_file_path = NULL,
  genehancer_gene_associations_scores_file_path = NULL,
  keep_only_double_elites = FALSE,
  skip_lifting_over = FALSE
)
}
\arguments{
\item{genehancer_data}{Data frame containing the GeneHancer data (can be imported from Excel or GFF files)}

\item{output_path}{Output path for the Genehancer reference file}

\item{genehancer_element_elite_status_file_path}{If available provide elite status data for GeneHancer elements. This argument takes in the file path, not the dataframe}

\item{genehancer_gene_associations_scores_file_path}{If available provide elite association data. This argument takes in the file path, not the dataframe}

\item{keep_only_double_elites}{Should Revana only consider double elite GeneHancer regions}

\item{skip_lifting_over}{By default data is lifted over from GRCh38 (hg38) to GRCh37 (hg19) coordinates. If this argument is set to true, original coordinates will be kept. Set this to TRUE if you plan on running Revana on GRCh38 data}
}
\description{
Prepare the GeneHancer reference file
}
