% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_motif_id_tf_gene_name_table.R
\name{prepare_motif_id_tf_gene_name_table_from_HOCOMOCO_annotation_tsv}
\alias{prepare_motif_id_tf_gene_name_table_from_HOCOMOCO_annotation_tsv}
\title{FIMO Motif Id - TF Gene Name Conversion Table File}
\usage{
prepare_motif_id_tf_gene_name_table_from_HOCOMOCO_annotation_tsv(
  annotation_file_path,
  output_path
)
}
\arguments{
\item{annotation_file_path}{path to the annotation file from HOCOMOCO}

\item{output_path}{output path for th created reference file}
}
\description{
FIMO Motif Id - TF Gene Name Conversion Table File
}
