% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_chipseq.R
\name{process_chipseq}
\alias{process_chipseq}
\title{Processessing of ChIP-Seq}
\usage{
process_chipseq(chipseq)
}
\arguments{
\item{chipseq}{tibble of ChIP-Seq data}
}
\description{
Processessing of ChIP-Seq
}
