% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_CNAs.R
\name{process_CNAs}
\alias{process_CNAs}
\title{Processing of Copy Number Aberations (CNAs)}
\usage{
process_CNAs(CNAs)
}
\arguments{
\item{CNAs}{tibble of somatic CNAs}
}
\description{
Processing of Copy Number Aberations (CNAs)
}
