% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_copy_numbers.R
\name{process_copy_numbers}
\alias{process_copy_numbers}
\title{Processing of Copy Number Data}
\usage{
process_copy_numbers(copy_numbers, sample_id)
}
\arguments{
\item{copy_numbers}{tibble of somatic copy number data}

\item{sample_id}{ID of the sample. Only used for adequate error messages}
}
\description{
Processing of Copy Number Data
}
