% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_genehancer_file.R
\name{process_genehancer}
\alias{process_genehancer}
\title{Processing of GeneHancer Reference File}
\usage{
process_genehancer(genehancer, gene_annotation)
}
\arguments{
\item{genehancer}{tibble of GeneHancers}

\item{gene_annotation}{tibble of GeneAnnotation}
}
\description{
Processing of GeneHancer Reference File
}
