% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_markers.R
\name{process_markers}
\alias{process_markers}
\title{Processessing of Single Nucleotide Markers}
\usage{
process_markers(markers, CNAs)
}
\arguments{
\item{markers}{tibble of markers}

\item{CNAs}{tibble of CNAs}
}
\description{
Processessing of Single Nucleotide Markers
}
