% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_somatic_SNVs.R
\name{process_somatic_SNVs_chunkwise}
\alias{process_somatic_SNVs_chunkwise}
\title{Processessing of somatic SNVs - chunkwise implementation}
\usage{
process_somatic_SNVs_chunkwise(
  somatic_SNVs,
  expression,
  motif_id_tf_gene_name_table,
  fimo_in_file_path_prefix,
  fimo_out_file_path_prefix,
  fimo_log_file_path_prefix,
  fimo_in_file_path_suffix,
  fimo_out_file_path_suffix,
  fimo_log_file_path_suffix,
  fimo_motif_ref_path,
  SNV_chunk_size = 10000,
  delete_FIMO_out_right_away = TRUE,
  reference_genome = NULL
)
}
\arguments{
\item{somatic_SNVs}{tibble of somatic SNVs}

\item{expression}{tibble of FPKM expression data}

\item{motif_id_tf_gene_name_table}{translation table from FIMO motifs to transcription factor gene names}

\item{fimo_in_file_path_prefix}{path to where should the FIMO input file be saved. Path is assembled like this PREFIX_CHUNK-NUMBER_SUFFIX}

\item{fimo_out_file_path_prefix}{path to where  FIMO should save its output file. Path is assembled like this PREFIX_CHUNK-NUMBER_SUFFIX}

\item{fimo_log_file_path_prefix}{path to where FIMO should save its log file. Path is assembled like this PREFIX_CHUNK-NUMBER_SUFFIX}

\item{fimo_in_file_path_suffix}{path to where should the FIMO input file be saved. Path is assembled like this PREFIX_CHUNK-NUMBER_SUFFIX}

\item{fimo_out_file_path_suffix}{path to where  FIMO should save its output file. Path is assembled like this PREFIX_CHUNK-NUMBER_SUFFIX}

\item{fimo_log_file_path_suffix}{path to where FIMO should save its log file. Path is assembled like this PREFIX_CHUNK-NUMBER_SUFFIX}

\item{fimo_motif_ref_path}{path to the FIMO motif reference file}

\item{SNV_chunk_size}{input size of the SNV chunk (number of SNV rows to be used for each chunk)}

\item{delete_FIMO_out_right_away}{should the FIMO data be deleted right after using it? TRUE: Eliminates lots of unnecessary output data}

\item{reference_genome}{Which genome should be used. Default (NULL) uses Human Genome build GRCh37 (hg19)}
}
\description{
Processessing of somatic SNVs - chunkwise implementation
}
