% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_SVs.R
\name{process_SVs}
\alias{process_SVs}
\title{Processessing of Structural Variants (SVs)}
\usage{
process_SVs(SVs)
}
\arguments{
\item{SVs}{tibble of somatic SNVs}
}
\description{
Processessing of Structural Variants (SVs)
}
