% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_per_sample.R
\name{run_step2_per_sample}
\alias{run_step2_per_sample}
\title{Run Step 2 of the Revana workflow on a cohort/subgroup of tumor sample data - Parallelized implementation}
\usage{
run_step2_per_sample(
  paths_file_path,
  output_dir,
  sample_id,
  gene_annotation_ref_file_path,
  gene_annotation_exons_ref_file_path,
  fimo_motif_ref_path = NULL,
  motif_id_tf_gene_name_table_path = NULL,
  run_tf_binding_site_analysis = TRUE,
  reference_genome = NULL,
  verbose = FALSE
)
}
\arguments{
\item{paths_file_path}{Path to the paths file. The paths file contains all relevant paths for all the samples to be included in the analysis. See the documentation for the exact format of this file.}

\item{output_dir}{Where are the analysis results to be stored}

\item{sample_id}{ID of the sample}

\item{gene_annotation_ref_file_path}{Path to the reference file for the gene annotation. See the documentation for the exact format of this file and how to obtain it.}

\item{gene_annotation_exons_ref_file_path}{Path to the reference file for the exon annotation. See the documentation for the exact format of this file and how to obtain it.}

\item{fimo_motif_ref_path}{Path to the reference file containing the FIMO motifs. See the documentation for the exact format of this file and how to obtain it.}

\item{motif_id_tf_gene_name_table_path}{Path to the reference for the TF gene name annotation of the FIMO motifs. See the documentation for the exact format of this file and how to obtain it.}

\item{run_tf_binding_site_analysis}{Should TF binding site analysis be conducted?}

\item{reference_genome}{Which genome should be used. Default (NULL) uses Human Genome build GRCh37 (hg19)}

\item{verbose}{should verbose logging be activated? (TRUE / FALSE)}
}
\description{
Run Step 2 of the Revana workflow on a cohort/subgroup of tumor sample data - Parallelized implementation
}
\details{
This function runs step 2 on a single sample.

For each included sample step 2 of the revana workflow ...
\itemize{
\item processes SNP marker input and calculates allele specific expression (ASE)
\item processes somatic SNV/InDel input
\item runs TF binding site analysis (if run_tf_binding_site_analysis == TRUE)
\item processes copy number input and calculates gene copy numbers and coverage ratios
\item processes expression input
\item processes copy number aberration input (CNAs)
\item processes structural variants (SVs)
}
}
