% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_per_cohort.R
\name{run_step3_per_cohort}
\alias{run_step3_per_cohort}
\title{Run Step 3 of the Revana workflow on a cohort/subgroup of tumor sample data}
\usage{
run_step3_per_cohort(
  paths_file_path,
  output_dir,
  gene_annotation_ref_file_path,
  TAD_file_path,
  genehancer_ref_file_path = NULL,
  chipseq_file_path = NULL,
  verbose = FALSE
)
}
\arguments{
\item{paths_file_path}{Path to the paths file. The paths file contains all relevant paths for all the samples to be included in the analysis. See the documentation for the exact format of this file.}

\item{output_dir}{Where are the analysis results to be stored}

\item{gene_annotation_ref_file_path}{Path to the reference file for the gene annotation. See the documentation for the exact format of this file and how to obtain it.}

\item{TAD_file_path}{Path to the TAD file. The TAD file contains genomic coordinates of Topologically Associated Domains (TADs). See the documentation for the exact format of this file.}

\item{genehancer_ref_file_path}{Path to the reference for GeneHancer. See the documentation for the exact format of this file and how to obtain it.}

\item{chipseq_file_path}{Path to the ChIP-Seq file. The ChIP-Seq file contains genomic coordinates of regulatory active regions and can be obtained from experimental data or online less specific online resources. This file is optional, although an empty dummy file has to be used, if no ChIP-Seq data is supplied. See the documentation for the exact format of this file.}

\item{verbose}{should verbose logging be activated? (TRUE / FALSE)}
}
\description{
Run Step 3 of the Revana workflow on a cohort/subgroup of tumor sample data
}
\details{
Step 3 of the revana workflow ...
\itemize{
\item creates the internal OHE reference
\item processes GeneHancer reference data
\item processes ChIPseq input data
\item links the ChIPseq regions to their respective TADs
}
}
