% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4_per_sample.R
\name{run_step4_per_cohort}
\alias{run_step4_per_cohort}
\title{Run Step 4 of the Revana workflow on a cohort/subgroup of tumor sample data - Parallelized implementation}
\usage{
run_step4_per_cohort(
  paths_file_path,
  output_dir,
  TAD_file_path,
  run_tf_binding_site_analysis = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{paths_file_path}{Path to the paths file. The paths file contains all relevant paths for all the samples to be included in the analysis. See the documentation for the exact format of this file.}

\item{output_dir}{Where are the analysis results to be stored}

\item{TAD_file_path}{Path to the TAD file. The TAD file contains genomic coordinates of Topologically Associated Domains (TADs). See the documentation for the exact format of this file.}

\item{run_tf_binding_site_analysis}{Should TF binding site analysis be conducted?}

\item{verbose}{should verbose logging be activated? (TRUE / FALSE)}
}
\description{
Run Step 4 of the Revana workflow on a cohort/subgroup of tumor sample data - Parallelized implementation
}
\details{
This implementation of step 4 runs the samples in parallel. This is the optimized and default implementation.

For each included sample step 4 of the revana workflow ...
\itemize{
\item calls cis-activation
\item adds the TADs to the genes
\item associates genes with regulatory regions from GeneHancer and ChIP-Seq data
\item matches somatic SNVs and InDels with genes and genehancers
\item matches somatic SNVs and InDels TF binding data with genes and genehancers
\item matches copy number aberrations (CNAs) with genes and genehancers
\item matches structural variants (SVs) with genes and genehancers
}
}
