% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step5_per_cohort.R
\name{run_step5_per_cohort}
\alias{run_step5_per_cohort}
\title{Run Step 5 of the Revana workflow on a cohort/subgroup of tumor sample data}
\usage{
run_step5_per_cohort(
  paths_file_path,
  output_dir,
  subgroup_name = "DEFAULT_GROUP",
  skip_merging_Rds_chunks = FALSE,
  run_tf_binding_site_analysis = TRUE
)
}
\arguments{
\item{paths_file_path}{Path to the paths file. The paths file contains all relevant paths for all the samples to be included in the analysis. See the documentation for the exact format of this file.}

\item{output_dir}{Where are the analysis results to be stored}

\item{subgroup_name}{Name of the subgroup. The name of the subgroup is used as reference in later created HTML reports}

\item{skip_merging_Rds_chunks}{Logical value. Should data chunks stored in .Rds-files be merged together. Default is TRUE - this step is required for regular report generation.}

\item{run_tf_binding_site_analysis}{Was TF binding site analysis conducted?}
}
\description{
Run Step 5 of the Revana workflow on a cohort/subgroup of tumor sample data
}
\details{
Step 5 of the revana workflow ...
\itemize{
\item merges .Rds file chunks (if skip_merging_Rds_chunks == TRUE)
\item creates results paths file
}
}
